[YS-290] refactor: 실험 개요 컴포넌트 로직 분리
